backend/requirements.txt
Flask
Flask-JWT-Extended
Flask-SQLAlchemy
Flask-CORS
psycopg2-binary
python-dotenv
marshmallow